requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
jinja2==3.1.6
openai-whisper==20250625
psutil==6.1.0
